backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lost-found-backend"
version = "1.0.0"
description = "Lost and Found Management System Backend API"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "redis>=5.0.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "celery>=5.3.6",
    "boto3>=1.34.34",
    "pillow>=10.2.0",
    "elasticsearch>=8.12.0",
    "python-dotenv>=1.0.1",
    "slowapi>=0.1.9",
    "hiredis>=2.3.2",
    "python-multipart>=0.0.6",
    "email-validator>=2.3.0",
    "argon2-cffi>=25.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.14",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["alembic*", "tests*"]

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
